backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
langchain==0.0.128
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pydantic==1.10.7
pytz==2023.3
PyYAML==6.0
requests==2.28.2
SQLAlchemy==1.4.47
sqlparse==0.4.3
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
